LayoutTests/fast/spatial-navigation/snav-offscreen-content-expected.txt

offscreen contentRight link

Down link
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
PASS successfullyParsed is true

TEST COMPLETE
